Make doc navigation easier
This adds a "back to top" button on the docs as well as
easy anchor links, so that users can use links to specific
sections of the docs.

The codeblock-label plugin will allow us to specify the
file name of code snippets, although it will render weirdly
when viewed on GitHub or in an external Markdown viewer.