Disable shrinkwrapping by default.
SDK app developers should not need to install internal dev
dependencies (e.g. Babel), so the status quo in which we distribute
the our SDKs with a shrinkwrap file that includes dev dependencies is
unacceptable.  However, internally we would still like to be able to
reproduce builds if needed, especially if packages or tools we depend
on become broken.

This change adds a helper script for enabling/disabling shrinkwrapped
installs, with the default state and the published state being
disabled. Flux internal SDK developers can run `npm run
obey-shrinkwrap` to cause `npm install` to use locked down
dependencies from the latest published SDK. The SDK build process
still includes the shrinkwrapping step so that we can always reproduce
the latest successful build.